+++ src/Tesseract/IResultRenderer.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Tesseract
+{
+    public interface IResultRenderer : IDisposable
+    {
+        IDisposable BeginDocument(string name);
+
+        void AddPage(Page page);
+    }
+}